Makefile.am: revert encoding of man page to UTF-8 (reverts effect of b3c8e4d)
Identified the original problem, which is that help2man needs to be built
with Locale::gettext (it is not enough merely to install the Perl module on
a system with a version of help2man built without it). Hence, do this for
macOS (where brew ships a gettext-less build of help2man).